feat: Update font settings in head.php and uno.config.ts
This commit includes changes in the font settings of the website. In
head.php, new font links are added and the body and code font-family
properties are updated. In uno.config.ts, the presetWebFonts function
is removed from the presets array. This change is part of an effort to
improve the website's typography and overall aesthetic.